requirements.txt
Flask==0.12.2

# Application server for both development and production.
gunicorn==19.7.1

# Python library to communicate with postgres db
psycopg2==2.7.3.2

#Flask extension to make working with SQLAlchemy easier
Flask-SQLAlchemy==2.3.1

# Forms.
#Flask-WTF==0.14.2
#WTForms-Components==0.10.3

Flask-WTF==0.9.5
WTForms-Components==0.9.7

# Build user login workflow
Flask-Login==0.3.2
#Flask-Login==0.4.1

# CLI.
Click==6.4

pytz
requests